project2/requirements.txt
requests
beautifulsoup4
pandas
gspread
gspread-dataframe
google-auth
google-auth-httplib2
google-auth-oauthlib
oauth2client
